+++ libflv/source/mpeg4-aac.c
@@ -74,7 +74,7 @@ int mpeg4_aac_adts_save(const struct mpeg4_aac_t* aac, size_t payload, uint8_t* 
 	data[4] = (uint8_t)(len >> 3);
 	data[5] = ((len & 0x07) << 5) | 0x1F;
 	data[6] = 0xFC /*| ((len / (1024 * aac->channels)) & 0x03)*/;
-	return len - payload;
+	return (int)(len - payload);
 }
 
 int mpeg4_aac_adts_frame_length(const uint8_t* data, size_t bytes)
