+++ libmov/source/mov-tfra.c
@@ -50,7 +50,7 @@ int mov_read_tfra(struct mov_t* mov, const struct mov_box_t* box)
 
 size_t mov_write_tfra(const struct mov_t* mov)
 {
-	size_t i, size;
+	uint32_t i, size;
 	const struct mov_track_t* track = mov->track;
 
 	size = 12/* full box */ + 12/* base */ + track->frag_count * 19/* index */;
