+++ librtp/payload/rtp-av1-unpack.c
@@ -109,7 +109,7 @@ static int rtp_decode_av1(void* p, const void* packet, int bytes)
 	}
 
 	pkt.payload = ptr;
-	pkt.payloadlen = pend - ptr;
+	pkt.payloadlen = (int)(pend - ptr);
 	rtp_payload_write(helper, &pkt);
 
 	if (pkt.rtp.m)
