+++ librtp/payload/rtp-vp8-unpack.c
@@ -153,7 +153,7 @@ static int rtp_decode_vp8(void* p, const void* packet, int bytes)
 	}
 
 	pkt.payload = ptr;
-	pkt.payloadlen = pend - ptr;
+	pkt.payloadlen = (int)(pend - ptr);
 	rtp_payload_write(helper, &pkt);
 
 	if (pkt.rtp.m)
